Change to way angles are set for reconstructed nuclei
KVReconstructedNucleus::GetAnglesFromStoppingDetector becomes
KVReconstructedNucleus::GetAnglesFromReconstructionTrajectory.

The former randomized particle direction either over surface of stopping detector,
or, if the detector just in front on the first trajectory associated with
the stopping detector (not the particle) had a smaller solid angle over that one.

Now we find the detector with the smallest solid angle on the entire
reconstruction trajectory associated with the particle and use that one.