Fix compilation issue with ROOT6.20: major change to build system
From now on we use whenever possible the cmake scripts & modules
from the installation of ROOT being used. As long as ROOT was built
with cmake these scripts are always available. Older versions may
have been built with configure/make, in this case the cmake modules
are missing. This method became obsolete with ROOT 6.14, and should
really only concern ROOT 5.

Therefore we now provide a minimal set of cmake modules (taken from
the latest version of 5.34) in case none are provided. We do not
check that we are effectively dealing with ROOT v5, in case it is
ROOT v6 built the old way, our build will fail (dictionaries generated
with rootcint instead of rootcling). But nobody should be building
ROOT6 this way (please?)