Backport from 1.10 of KVGemini & KVSimReader_HIPSE bugfixes
In KVGemini, add now rotational energy when initializing a CNucleus,
following K.Mazurek and M. Ciemala suggestions so if J>0, E* is
never 0 in CNucleus but follow the yrast line. It Could induce some
problems for very exotic nuclei since CYrast::getYrast() is defined only
for a defined range of isotopes per element.

This is because GEMINI++ expects excitation energy of nucleus to contain
both thermal & rotation (yrast) contributions (the rotational part is
subtracted in the code before calculating decay widths etc.)

In KVSimReader_HIPSE, the excitation energy read from output file is in
MeV/nucleon; this was not taken into account before, now it is corrected.